requirements.txt
google-adk
fastapi==0.121.1
uvicorn